Extracted RN2483 logic into its own class that is decoupled from the process message queues and logic.
